# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  avatarColor: string;
  createdAt: string;
}

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name: string;
}

export interface UserSummary {
  totalLogs: number;
  /** @nullable */
  averageOverall: number | null;
  longestStreakDays: number;
  /** @nullable */
  favoriteSpot: string | null;
  groupCount: number;
}

export interface Group {
  id: string;
  name: string;
  inviteCode: string;
  memberCount: number;
  createdAt: string;
}

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name: string;
  userId: string;
}

export interface GroupJoinInput {
  inviteCode: string;
  userId: string;
}

export interface GroupMember {
  id: string;
  name: string;
  avatarColor: string;
  joinedAt: string;
}

export interface GroupDetail {
  id: string;
  name: string;
  inviteCode: string;
  createdAt: string;
  members: GroupMember[];
}

export interface LeaderboardEntry {
  userId: string;
  name: string;
  avatarColor: string;
  totalLogs: number;
  /** @nullable */
  averageOverall: number | null;
}

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  speed: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  comfort: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  privacy: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  ambiance: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  relief: number;
}

export interface PoopLogInput {
  userId: string;
  /** @nullable */
  groupId?: string | null;
  lat: number;
  lng: number;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  notes?: string | null;
  ratings: RatingInput;
}

export interface PoopLog {
  id: string;
  userId: string;
  userName: string;
  avatarColor: string;
  /** @nullable */
  groupId: string | null;
  lat: number;
  lng: number;
  /** @nullable */
  locationName: string | null;
  /** @nullable */
  notes: string | null;
  ratings: RatingInput;
  overallScore: number;
  createdAt: string;
}

export type ListGroupsParams = {
userId: string;
};

export type ListLogsParams = {
userId: string;
};