# lib/api-zod/src/generated/types/group.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Group {
  id: string;
  name: string;
  inviteCode: string;
  memberCount: number;
  createdAt: Date;
}