# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name: string;
  userId: string;
}