# lib/api-zod/src/generated/types/groupJoinInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GroupJoinInput {
  inviteCode: string;
  userId: string;
}