# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  userId: string;
  name: string;
  avatarColor: string;
  totalLogs: number;
  /** @nullable */
  averageOverall: number | null;
}