# lib/api-zod/src/generated/types/poopLog.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RatingInput } from './ratingInput';

export interface PoopLog {
  id: string;
  userId: string;
  userName: string;
  avatarColor: string;
  /** @nullable */
  groupId: string | null;
  lat: number;
  lng: number;
  /** @nullable */
  locationName: string | null;
  /** @nullable */
  notes: string | null;
  ratings: RatingInput;
  overallScore: number;
  createdAt: Date;
}