# lib/api-zod/src/generated/types/poopLogInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RatingInput } from './ratingInput';

export interface PoopLogInput {
  userId: string;
  /** @nullable */
  groupId?: string | null;
  lat: number;
  lng: number;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  notes?: string | null;
  ratings: RatingInput;
}