# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  speed: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  comfort: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  privacy: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  ambiance: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  relief: number;
}