# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserSummary {
  totalLogs: number;
  /** @nullable */
  averageOverall: number | null;
  longestStreakDays: number;
  /** @nullable */
  favoriteSpot: string | null;
  groupCount: number;
}